Added FM, moved FilterBank to utils
- A whole file for one function seemed unnecessary
