+++ moabb/pipelines/utils.py
@@ -38,23 +38,30 @@ def create_pipeline_from_config(config):
 
 
 class FilterBank(BaseEstimator, TransformerMixin):
-    """Class that applies a given pipeline to multiple frequency bands from
-    MotorImageryMultiPass.
+    """Apply a given indentical pipeline over a bank of filter.
 
-    TODO: should this be forced to be a transformer?
+    The pipeline provided with the constrictor will be appield on the 4th
+    axis of the input data. This pipeline should be used with a FilterBank
+    paradigm.
+
+    This can be used to build a filterbank CSP, for example::
+
+        pipeline = make_pipeline(FilterBank(estimator=CSP()), LDA())
+
+    Parameters
+    ----------
+    estimator: sklean Estimator
+        the sklearn pipeline to apply on each band of the filter bank.
+    flatten: bool (True)
+        If True, output of each band are concatenated together on the feature
+        axis. if False, output are stacked.
     """
 
     def __init__(self, estimator, flatten=True):
-        """
-        estimator: BaseEstimator instance, applied on the 4th axis
-        flatten: bool, whether to flatten output of estimator
-        """
-        assert issubclass(type(estimator), BaseEstimator)
         self.estimator = estimator
         self.flatten = flatten
 
     def fit(self, X, y=None):
-        """fit"""
         assert X.ndim == 4
         self.models = [deepcopy(self.estimator).fit(X[..., i], y)
                        for i in range(X.shape[-1])]
@@ -63,14 +70,14 @@ class FilterBank(BaseEstimator, TransformerMixin):
     def transform(self, X):
         assert X.ndim == 4
         out = [self.models[i].transform(X[..., i]) for i in range(X.shape[-1])]
-        assert out[0].ndim == 2, 'output is of dim {}'.format(out[0].ndim)
+        assert out[0].ndim == 2, ("Each band must return a two dimensional "
+                                  f" matrix, currently have {out[0].ndim}")
         if self.flatten:
             return np.concatenate(out, axis=1)
         else:
             return np.stack(out, axis=2)
 
     def __repr__(self):
-        est = self.estimator.get_params()
         return '{}(estimator={}, flatten={})'.format(type(self).__name__,
-                                                     est,
+                                                     self.estimator.get_params(), # flake8: noqa
                                                      self.flatten)
